+++ packages/bundler/src/index.ts
@@ -1,4 +1,6 @@
-import { createAzionESBuildConfig } from './bundlers/esbuild/esbuild';
-import { createAzionWebpackConfig } from './bundlers/webpack/webpack';
+import { createAzionESBuildConfig, executeESBuildBuild } from './bundlers/esbuild/esbuild';
+import { createAzionWebpackConfig, executeWebpackBuild } from './bundlers/webpack/webpack';
 
-export { createAzionESBuildConfig, createAzionWebpackConfig };
+export { createAzionESBuildConfig, createAzionWebpackConfig, executeESBuildBuild, executeWebpackBuild };
+
+export * from './types/bundler';