+++ packages/config/src/configProcessor/processStrategy/implementations/domainProcessConfigStrategy.ts
@@ -9,8 +9,8 @@ import ProcessConfigStrategy from '../processConfigStrategy';
 class DomainProcessConfigStrategy extends ProcessConfigStrategy {
   transformToManifest(config: AzionConfig) {
     const domain = config?.domain;
-    if (!domain) {
-      return {};
+    if (!domain || Object.keys(domain).length === 0) {
+      return;
     }
     if (
       domain.digitalCertificateId &&
@@ -35,12 +35,12 @@ class DomainProcessConfigStrategy extends ProcessConfigStrategy {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const domainSetting: any = {
       name: domain.name,
-      cname_access_only: domain.cnameAccessOnly || false,
+      cname_access_only: domain.cnameAccessOnly ?? false,
       cnames: domain.cnames || [],
       digital_certificate_id: domain.digitalCertificateId || null,
       edge_application_id: domain.edgeApplicationId || null,
       edge_firewall_id: domain.edgeFirewallId || null,
-      active: true,
+      is_active: domain.active === undefined ? true : domain.active,
     };
     if (domain.mtls) {
       domainSetting.is_mtls_enabled = true;
@@ -55,7 +55,10 @@ class DomainProcessConfigStrategy extends ProcessConfigStrategy {
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   transformToConfig(payload: any, transformedPayload: AzionConfig) {
-    const domain = payload.domain || {};
+    const domain = payload.domain;
+    if (!domain || Object.keys(domain).length === 0) {
+      return;
+    }
     transformedPayload.domain = {
       name: domain.name,
       cnameAccessOnly: domain.cname_access_only,
@@ -63,11 +66,14 @@ class DomainProcessConfigStrategy extends ProcessConfigStrategy {
       digitalCertificateId: domain.digital_certificate_id,
       edgeApplicationId: domain.edge_application_id,
       edgeFirewallId: domain.edge_firewall_id,
-      mtls: {
-        verification: domain.mtls_verification,
-        trustedCaCertificateId: domain.mtls_trusted_ca_certificate_id,
-        crlList: domain.crl_list,
-      },
+      active: domain.is_active === undefined ? true : domain.is_active,
+      mtls: domain.mtls_verification
+        ? {
+            verification: domain.mtls_verification,
+            trustedCaCertificateId: domain.mtls_trusted_ca_certificate_id,
+            crlList: domain.crl_list,
+          }
+        : undefined,
     };
     return transformedPayload.domain;
   }