+++ packages/presets/src/index.ts
@@ -11,6 +11,7 @@ import JavaScript from './presets/javascript/index';
 import Jekyll from './presets/jekyll/index';
 import Next from './presets/next/index';
 import Nuxt from './presets/nuxt/index';
+import Preact from './presets/preact/index';
 import Qwik from './presets/qwik/index';
 import React from './presets/react/index';
 import RustWasm from './presets/rustwasm/index';
@@ -33,6 +34,7 @@ const presets = {
   Jekyll,
   Next,
   Nuxt,
+  Preact,
   Qwik,
   React,
   RustWasm,
@@ -56,6 +58,7 @@ export {
   Jekyll,
   Next,
   Nuxt,
+  Preact,
   Qwik,
   React,
   RustWasm,
@@ -66,3 +69,5 @@ export {
 };
 
 export default presets;
+
+export * from './types';