+++ packages/presets/src/presets/vue/metadata.ts
@@ -1,0 +1,5 @@
+import type { PresetMetadata } from 'azion/presets';
+
+export const metadata: PresetMetadata = {
+  name: 'vue',
+};