+++ packages/presets/src/types.ts
@@ -1,8 +1,15 @@
+import type { AzionBuild } from 'azion/config';
 import { AzionConfig } from 'azion/config';
+import { FetchEvent } from 'azion/types';
 
-export interface AzionBuildPreset {
+export type PresetMetadata = {
+  name: string;
+};
+
+export type AzionBuildPreset = {
   config: AzionConfig;
-  handler?: () => void;
-  prebuild?: () => void;
-  postbuild?: () => void;
-}
+  handler: (event: FetchEvent) => Promise<Response>;
+  prebuild?: (context: AzionBuild) => Promise<void>;
+  postbuild?: (context: AzionBuild) => Promise<void>;
+  metadata: PresetMetadata;
+};